Moved unrelated function from parser to new file
